Started work on new public device API
Getting a device by id will now always return an object. Internally the
public device is still a proxy but is now backed by an object that may
delegate to the local/remote device instance if it exists.

As part of this a new WeakRefMap has been created to keep track of if we
need to keep a public device alive or not. It will be kept alive ss long
as something is holding a reference to the object or it has a listener.